requirements.txt
# Core
Django==5.2.3
djangorestframework==3.15.2
django-environ==0.11.2

# Auth
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Filters & CORS
django-filter==24.3
django-cors-headers==4.4.0
